+++ src/types/MTGCard.ts
@@ -27,6 +27,22 @@ export async function getCardsForSet(setId: string): Promise<MTGCard[]> {
     return _getCardsForSet(setId);
 }
 
+export async function getCardsByQuery(query: string, setId: string | undefined): Promise<MTGCard[]> {
+    const finalSetId = setId && !query.match(/\be:[a-z0-9]{3,}/i) ? setId : undefined;
+    const params = {
+        order: 'set',
+        q: finalSetId ? `e:${finalSetId} ${query}` : query,
+        unique: 'prints'
+    };
+    try {
+        const scryfallCards = await fetchDataPaginated<ScryFallCard>(cardsSearchEndpoint, params, 500);
+        return scryfallCards.map(fromScryfallCard);
+    } catch (error) {
+        console.error('Error fetching cards by query:', error);
+        return [];
+    }
+}
+
 async function _getCardsForSet(setId: string): Promise<MTGCard[]> {
     const params = {
         order: 'set',
@@ -34,16 +50,7 @@ async function _getCardsForSet(setId: string): Promise<MTGCard[]> {
         unique: 'prints'
     };
     const scryfallCards = await fetchDataPaginated<ScryFallCard>(cardsSearchEndpoint, params);
-    const cards = scryfallCards.map(card => ({
-        id: card.id,
-        name: card.name,
-        setId: card.set,
-        series: card.set_name,
-        collectorNumber: card.collector_number,
-        imageUrl: card.image_uris?.normal || card.card_faces?.[0]?.image_uris?.normal || '',
-        color: card.color_identity?.join('') || 'C',
-        rarity: card.rarity,
-    }));
+    const cards = scryfallCards.map(fromScryfallCard);
     cache.bySet.set(setId, cards);
     return cards;
 }
@@ -52,4 +59,17 @@ export async function getCard(setId: string, collectorNumber: string): Promise<M
     const cards = await getCardsForSet(setId);
     const card = cards.find(c => c.collectorNumber === collectorNumber);
     return card || null;
+}
+
+function fromScryfallCard(card: ScryFallCard): MTGCard {
+    return {
+        id: card.id,
+        name: card.name,
+        setId: card.set,
+        series: card.set_name,
+        collectorNumber: card.collector_number,
+        imageUrl: card.image_uris?.normal || card.card_faces?.[0]?.image_uris?.normal || '',
+        color: card.color_identity?.join('') || 'C',
+        rarity: card.rarity,
+    };
 }